main: get font paths using fontconfig and make Makefile more packager friendly
By using fontconfig the program can find a font without hardcoding paths
which only work on some distros, but not others.

In addition, the Makefile is changed to allow for easier building and
changing where the package is installed.

Lastly, add a guix.scm file for easier development and as a proof of
concept on how to package simplefuckingtimer for distros.